deepfake-terminal-ui/api/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0

# PyTorch & Model
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
